+++ src/test/groovy/com/github/choonchernlim/security/adfs/saml2/SpringSecurityAdfsSaml2ExceptionSpec.groovy
@@ -1,0 +1,15 @@
+package com.github.choonchernlim.security.adfs.saml2
+
+import spock.lang.Specification
+
+class SpringSecurityAdfsSaml2ExceptionSpec extends Specification {
+
+    def "exception"() {
+        when:
+        def exception = new SpringSecurityAdfsSaml2Exception('test')
+
+        then:
+        exception instanceof RuntimeException
+        exception.message == 'test'
+    }
+}